Refactor Solana token resolvers and update import paths
- Moved Solana token-related functions into a dedicated `resolvers` directory for better organization.
- Updated import paths across multiple files to reflect the new structure, ensuring consistency and maintainability.
- Introduced new resolver functions for fetching Solana tokens and balances, enhancing modularity and clarity in the codebase.
