+++ clients/extension/src/utils/tx/getStandardTx.ts
@@ -7,7 +7,7 @@ import { base64 } from '@coral-xyz/anchor/dist/cjs/utils/bytes'
 import { Chain } from '@core/chain/Chain'
 import { getCosmosChainByChainId } from '@core/chain/chains/cosmos/chainInfo'
 import { chainFeeCoin } from '@core/chain/coin/chainFeeCoin'
-import { getSolanaToken } from '@core/chain/coin/find/solana/getSolanaToken'
+import { getSolanaToken } from '@core/chain/coin/find/resolvers/solana/getSolanaToken'
 import { match } from '@lib/utils/match'
 import { TW } from '@trustwallet/wallet-core'
 import { bech32 } from 'bech32'
