+++ core/ui/vault/chain/coin/finder/queries/useCoinFinderQuery.ts
@@ -2,7 +2,7 @@ import { getChainKind } from '@core/chain/ChainKind'
 import { AccountCoin } from '@core/chain/coin/AccountCoin'
 import { findCoins } from '@core/chain/coin/find'
 import { coinFinderChainKinds } from '@core/chain/coin/find/CoinFinderChainKind'
-import { FindCoinsResolverInput } from '@core/chain/coin/find/FindCoinsResolver'
+import { FindCoinsResolverInput } from '@core/chain/coin/find/resolver'
 import { useCurrentVaultAddresses } from '@core/ui/vault/state/currentVaultCoins'
 import { useQueriesToEagerQuery } from '@lib/ui/query/hooks/useQueriesToEagerQuery'
 import { noRefetchQueryOptions } from '@lib/ui/query/utils/options'
